+++ inc/api/FTDI_Device_close.h
@@ -1,8 +1,0 @@
-#ifndef FTDI_Device_close_h
-#define FTDI_Device_close_h
-
-#include <node_api.h>
-
-napi_value device_close(napi_env env, napi_callback_info info);
-
-#endif