+++ inc/api/FTDI_Device_purge.h
@@ -1,8 +1,0 @@
-#ifndef FTDI_Device_purge_h
-#define FTDI_Device_purge_h
-
-#include <node_api.h>
-
-napi_value device_purge(napi_env env, napi_callback_info info);
-
-#endif