+++ inc/api/FTDI_Device_read.h
@@ -1,8 +1,0 @@
-#ifndef FTDI_Device_read_h
-#define FTDI_Device_read_h
-
-#include <node_api.h>
-
-napi_value device_read(napi_env env, napi_callback_info info);
-
-#endif