+++ inc/api/FTDI_Device_setBaudRate.h
@@ -1,8 +1,0 @@
-#ifndef FTDI_Device_setBaudRate_h
-#define FTDI_Device_setBaudRate_h
-
-#include <node_api.h>
-
-napi_value device_setBaudRate(napi_env env, napi_callback_info info);
-
-#endif