+++ inc/api/FTDI_Device_write.h
@@ -1,8 +1,0 @@
-#ifndef FTDI_Device_write_h
-#define FTDI_Device_write_h
-
-#include <node_api.h>
-
-napi_value device_write(napi_env env, napi_callback_info info);
-
-#endif