+++ inc/api/FTDI_Device_writeEE.h
@@ -1,8 +1,0 @@
-#ifndef FTDI_Device_writeEE_h
-#define FTDI_Device_writeEE_h
-
-#include <node_api.h>
-
-napi_value device_writeEE(napi_env env, napi_callback_info info);
-
-#endif