+++ main.d.ts
@@ -1,4 +1,6 @@
 declare class FTDI_DeviceInfo {
+  /** @hidden */ constructor();
+
   /** The device serial number */
   readonly serial_number: string;
   /** The device description */
@@ -20,6 +22,8 @@ declare class FTDI_DeviceInfo {
 }
 
 declare class FTDI_Device {
+  /** @hidden */ constructor();
+  
   /** The device serial number */
   readonly serial_number: string;
   /** Getter to device connection status:
@@ -80,9 +84,35 @@ declare class FTDI_Device {
   /** Getter to the device driver version */
   get driver_version(): string;
 
+  /** Getter to the current value of the latency timer (supported devices only)
+   * 
+   * In the FT8U232AM and FT8U245AM devices, the receive buffer timeout that is used to flush remaining
+   * data from the receive buffer was fixed at 16 ms. In all other FTDI devices, this timeout is programmable
+   * and can be set at 1 ms intervals between 2ms and 255 ms. This allows the device to be better optimized
+   * for protocols requiring faster response times from short data packets. */
+   get latency_timer(): number;
+
+  /** Getter to the instantaneous value of the data bus
+   * 
+   * For a description of available bit modes for the FT232R, see the application note "Bit Bang Modes for the
+   * FT232R and FT245R".
+   * 
+   * For a description of available bit modes for the FT2232, see the application note "Bit Mode Functions for
+   * the FT2232".
+   * 
+   * For a description of bit bang modes for the FT232B and FT245B, see the application note
+   * "FT232B/FT245B Bit Bang Mode".
+   * 
+   * For a description of bit modes supported by the FT4232H and FT2232H devices, please see the IC data
+   * sheets.
+   * 
+   * These application notes are available for download from the FTDI website. */
+  get bit_mode(): number;
+
   /** Close the device
    * 
-   * IMPORTANT: Device object is considered dead after this function has been called. Any call to a function or getter of this object will throw an error. Device should be reopened using the {@link openDevice} function that will generate a new {@link FTDI_Device} object.
+   * > **Warning**
+   * > IMPORTANT: Device object is considered dead after this function has been called. Any call to a function or getter of this object will throw an error. Device should be reopened using the {@link openDevice} function that will generate a new {@link FTDI_Device} object.
    */
   close(): Promise<void>;
 
@@ -105,7 +135,7 @@ declare class FTDI_Device {
 
   /** Sets the baud rate for the device
    * @param baud_rate The baud rate to set to the device (bits per second)
-  */
+   */
   setBaudRate(baud_rate: number): void;
 
   /** Sets the data characteristics for the device
@@ -163,16 +193,51 @@ declare class FTDI_Device {
 
   /** Erases the device EEPROM */
   eraseEE(): Promise<void>;
+
+  /** Sets the latency timer value (supported devices only)
+   * @param timer Latency timer value, in milliseconds. Valid range is 2 to 255.
+   */
+  setLatencyTimer(timer: number): void;
+
+  /** Enables different chip modes (puts the device in a mode other than the default UART or FIFO mode).
+    * @param mask Bit mode mask
+    * Sets up which bits are inputs and which are outputs. 
+    * A bit value of 0 sets the corresponding pin to an input, a bit value of 1 sets the corresponding pin to an output. 
+    * In the case of CBUS Bit Bang, the upper nibble of this value controls which pins are inputs and outputs, while the lower nibble controls which of the outputs are high and low.
+    * @param bit_mode Mode value. Can be one of the following:
+    * For FT232H devices, valid values are FT_BIT_MODE_RESET, FT_BIT_MODE_ASYNC_BITBANG, FT_BIT_MODE_MPSSE, FT_BIT_MODE_SYNC_BITBANG, FT_BIT_MODE_CBUS_BITBANG, FT_BIT_MODE_MCU_HOST, FT_BIT_MODE_FAST_SERIAL, FT_BIT_MODE_SYNC_FIFO.
+    * For FT2232H devices, valid values are FT_BIT_MODE_RESET, FT_BIT_MODE_ASYNC_BITBANG, FT_BIT_MODE_MPSSE, FT_BIT_MODE_SYNC_BITBANG, FT_BIT_MODE_MCU_HOST, FT_BIT_MODE_FAST_SERIAL, FT_BIT_MODE_SYNC_FIFO.
+    * For FT4232H devices, valid values are FT_BIT_MODE_RESET, FT_BIT_MODE_ASYNC_BITBANG, FT_BIT_MODE_MPSSE, FT_BIT_MODE_SYNC_BITBANG.
+    * For FT232R devices, valid values are FT_BIT_MODE_RESET, FT_BIT_MODE_ASYNC_BITBANG, FT_BIT_MODE_SYNC_BITBANG, FT_BIT_MODE_CBUS_BITBANG.
+    * For FT245R devices, valid values are FT_BIT_MODE_RESET, FT_BIT_MODE_ASYNC_BITBANG, FT_BIT_MODE_SYNC_BITBANG.
+    * For FT2232 devices, valid values are FT_BIT_MODE_RESET, FT_BIT_MODE_ASYNC_BITBANG, FT_BIT_MODE_MPSSE, FT_BIT_MODE_SYNC_BITBANG, FT_BIT_MODE_MCU_HOST, FT_BIT_MODE_FAST_SERIAL.
+    * For FT232B and FT245B devices, valid values are FT_BIT_MODE_RESET, FT_BIT_MODE_ASYNC_BITBANG.
+    */
+  setBitMode(mask: number, bit_mode: number): void;
+
+  /** Sets the USB request transfer size
+   * 
+   * This function can be used to change the transfer sizes from the default transfer size of 4096 bytes to
+   * better suit the application requirements. Transfer sizes must be set to a multiple of 64 bytes between 64
+   * bytes and 64k bytes.
+   * When FT_SetUSBParameters is called, the change comes into effect immediately and any data that was
+   * held in the driver at the time of the change is lost.
+   * Note that, at present, only `in_transfer_size` is supported by the FTDI vendor driver.
+   *
+   * @param in_transfer_size Transfer size in bits for USB IN request
+   * @param out_transfer_size Transfer size in bits for USB OUT request (unsupported yet)
+   */
+  setUSBParameters(in_transfer_size: number, out_transfer_size: number): void;
 }
 
 /** Get the list of FTDI devices connected to the system */
 export function getDeviceInfoList(): Promise<FTDI_DeviceInfo[]>;
 
-/** Try to open a device from its serial number
- * @param serial_number Device serial number
+/** Try to open a device
+ * @param identifier Device serial number (string), or an object containing either its `serial_number` (string), its `usb_loc_id` (number) or its `description` (string) property
  * @returns Device object
  */
-export function openDevice(serial_number: string): Promise<FTDI_Device>;
+export function openDevice(identifier: string | { serial_number?: string, usb_loc_id?: number, description?: string }): Promise<FTDI_Device>;
 
 /** This allows the driver to load only for the specified VID and PID combination.
  * 
@@ -204,3 +269,11 @@ export const FT_FLOW_DTR_DSR: number;
 export const FT_FLOW_XON_XOFF: number;
 export const FT_PURGE_RX: number;
 export const FT_PURGE_TX: number;
+export const FT_BITMODE_RESET: number;
+export const FT_BITMODE_ASYNC_BITBANG: number;
+export const FT_BITMODE_MPSSE: number;
+export const FT_BITMODE_SYNC_BITBANG: number;
+export const FT_BITMODE_MCU_HOST: number;
+export const FT_BITMODE_FAST_SERIAL: number;
+export const FT_BITMODE_CBUS_BITBANG: number;
+export const FT_BITMODE_SYNC_FIFO: number;