+++ src/api/FTDI_Device_eraseEE.c
@@ -1,6 +1,5 @@
 #include <stdlib.h>
 
-#include "api/FTDI_Device_eraseEE.h"
 #include "api/FTDI_Device.h"
 #include "ftd2xx.h"
 #include "utils.h"
@@ -73,7 +72,7 @@ napi_value device_eraseEE(napi_env env, napi_callback_info info) {
 
   // Allocate memory for async instance data structure
   async_data_t* async_data = malloc(sizeof(async_data_t));
-  if(utils_check(async_data == NULL, "Malloc failed", "malloc")) return NULL;
+  if(utils_check(async_data == NULL, "Malloc failed", ERR_MALLOC)) return NULL;
 
   // Get the class instance data containing FTDI device handle
   utils_check(napi_unwrap(env, this_arg, (void**)&(async_data->instance_data)));