+++ src/api/FTDI_Device_setTimeouts.c
@@ -1,4 +1,3 @@
-#include "api/FTDI_Device_setTimeouts.h"
 #include "api/FTDI_Device.h"
 #include "ftd2xx.h"
 #include "utils.h"
@@ -9,22 +8,22 @@ napi_value device_setTimeouts(napi_env env, napi_callback_info info) {
   size_t argc = NB_ARGS; // size of the argv buffer
   napi_value this_arg, argv[NB_ARGS];
   utils_check(napi_get_cb_info(env, info, &argc, argv, &this_arg, NULL));
-  if(utils_check(argc < NB_ARGS, "Missing argument", "missarg")) return NULL;
+  if(utils_check(argc < NB_ARGS, "Missing argument", ERR_MISSARG)) return NULL;
 
   // Get the class instance data containing FTDI device handle
   device_instance_data_t* instance_data;
   utils_check(napi_unwrap(env, this_arg, (void**)(&instance_data)));
 
   // Check the device is open if its handle is still there
-  if(utils_check(instance_data->ftHandle == NULL, "Dead device object", "deadobj")) return NULL;
+  if(utils_check(instance_data->ftHandle == NULL, "Dead device object", ERR_DEADOBJ)) return NULL;
 
   // Check arguments, and convert JavaScript values to C values
   uint32_t read_timeout, write_timeout;
-  if(utils_check(napi_get_value_uint32(env, argv[0], &read_timeout), "Read timeout must be a number", "wrongarg")) return NULL;
-  if(utils_check(napi_get_value_uint32(env, argv[1], &write_timeout), "Write timeout must be a number", "wrongarg")) return NULL;
+  if(utils_check(napi_get_value_uint32(env, argv[0], &read_timeout), "Read timeout must be a number", ERR_WRONGARG)) return NULL;
+  if(utils_check(napi_get_value_uint32(env, argv[1], &write_timeout), "Write timeout must be a number", ERR_WRONGARG)) return NULL;
 
   // Update FTDI device
-  utils_check(FT_SetTimeouts(instance_data->ftHandle, read_timeout, write_timeout));
+  utils_check(FT_SetTimeouts(instance_data->ftHandle, (ULONG)read_timeout, (ULONG)write_timeout));
 
   return NULL;
 }