+++ src/api/FTDI_Device_setUSBParameters.c
@@ -2,9 +2,9 @@
 #include "ftd2xx.h"
 #include "utils.h"
 
-napi_value device_purge(napi_env env, napi_callback_info info) {
+napi_value device_setUSBParameters(napi_env env, napi_callback_info info) {
   // Get JavaScript `this` corresponding to this instance of the class and `argc`/`argv` passed to the function
-  #define NB_ARGS 1 // number of expected arguments
+  #define NB_ARGS 2 // number of expected arguments
   size_t argc = NB_ARGS; // size of the argv buffer
   napi_value this_arg, argv[NB_ARGS];
   utils_check(napi_get_cb_info(env, info, &argc, argv, &this_arg, NULL));
@@ -18,12 +18,12 @@ napi_value device_purge(napi_env env, napi_callback_info info) {
   if(utils_check(instance_data->ftHandle == NULL, "Dead device object", ERR_DEADOBJ)) return NULL;
 
   // Check arguments, and convert JavaScript values to C values
-  uint32_t mask;
-  if(utils_check(((napi_get_value_uint32(env, argv[0], &mask) != napi_ok) || (mask & 0b11) == 0),
-      "Mask must be a combination of of FT_PURGE_RX and FT_PURGE_TX", ERR_WRONGARG)) return NULL;
+  uint32_t in_transfer_size, out_transfer_size;
+  if(utils_check(napi_get_value_uint32(env, argv[0], &in_transfer_size), "Transfer size for USB IN request must be a number", ERR_WRONGARG)) return NULL;
+  if(utils_check(napi_get_value_uint32(env, argv[1], &out_transfer_size), "Transfer size for USB OUT request must be a number", ERR_WRONGARG)) return NULL;
 
   // Update FTDI device
-  utils_check(FT_Purge(instance_data->ftHandle, mask));
+  utils_check(FT_SetUSBParameters(instance_data->ftHandle, (ULONG)in_transfer_size, (ULONG)out_transfer_size));
 
   return NULL;
 }