+++ src/api/FTDI_Device_write.c
@@ -1,7 +1,6 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "api/FTDI_Device_write.h"
 #include "api/FTDI_Device.h"
 #include "ftd2xx.h"
 #include "utils.h"
@@ -42,7 +41,7 @@ static void complete_callback(napi_env env, napi_status status, void* data) {
   async_data_t* async_data = (async_data_t*) data;
 
   // Manage FTDI error if any
-  utils_check(async_data->ftStatus == FT_IO_ERROR, "USB was lost during write operation", "usblost");
+  utils_check(async_data->ftStatus == FT_IO_ERROR, "USB was lost during write operation", ERR_USBLOST);
   utils_check(FT_|async_data->ftStatus); // manage other errors
 
   // Resolve the JavaScript `Promise`:
@@ -79,16 +78,16 @@ napi_value device_write(napi_env env, napi_callback_info info) {
   size_t argc = NB_ARGS; // size of the argv buffer
   napi_value this_arg, argv[NB_ARGS];
   utils_check(napi_get_cb_info(env, info, &argc, argv, &this_arg, NULL));
-  if(utils_check(argc < NB_ARGS, "Missing argument", "missarg")) return NULL;
+  if(utils_check(argc < NB_ARGS, "Missing argument", ERR_MISSARG)) return NULL;
 
   // Check that the data argument is a TypedArray
   bool is_typedarray;
   utils_check(napi_is_typedarray(env, argv[0], &is_typedarray));
-  if(utils_check(is_typedarray == false, "The data to write must be a TypedArray", "wrongarg")) return NULL;
+  if(utils_check(is_typedarray == false, "The data to write must be a TypedArray", ERR_WRONGARG)) return NULL;
 
   // Allocate memory for async instance data structure
   async_data_t* async_data = malloc(sizeof(async_data_t));
-  if(utils_check(async_data == NULL, "Malloc failed", "malloc")) return NULL;
+  if(utils_check(async_data == NULL, "Malloc failed", ERR_MALLOC)) return NULL;
 
   // Allocate memory for TX buffer
   void* data;
@@ -98,7 +97,7 @@ napi_value device_write(napi_env env, napi_callback_info info) {
   utils_check(napi_get_arraybuffer_info(env, array_buffer, &data, &byte_length));
   byte_length -= byte_offset;
   async_data->tx_buffer = malloc(byte_length);
-  if(utils_check(async_data->tx_buffer == NULL, "Malloc failed", "malloc")) return NULL;
+  if(utils_check(async_data->tx_buffer == NULL, "Malloc failed", ERR_MALLOC)) return NULL;
 
   // Copy TX buffer
   memcpy(async_data->tx_buffer, (uint8_t*)data + byte_offset, byte_length);