+++ test.js
@@ -1,21 +1,28 @@
 const repl = require("repl");
+const fs = require("fs");
 
 // Load FTDI library
-let FTDI;
-try { FTDI = require("./build/Debug/ftdi-d2xx.node"); } // Try loading the Debug build, if any
-catch (e) { FTDI = require("./main.js"); } // Else load the main.js file that loads the right Release build
+fs.stat("./build/Debug/ftdi-d2xx.node", async (err) => { // Check if the Debug build exists
+  let FTDI;
+  if(!err) {
+    console.log("test.js: Loading the Debug Build...")
+    FTDI = require("./build/Debug/ftdi-d2xx.node"); // Try loading the Debug build, if any
+  }
+  else {
+    console.log("test.js: Warning! Debug Build was not found under /build/Debug/ftdi-d2xx.node, loading the Release Build...")
+    FTDI = require("./main.js"); // Else load the main.js file that loads the right Release build
+  }
 
-// Print some testing
-(async function () {
+  // Print some testing
   try {
     const devices = await FTDI.getDeviceInfoList();
-    console.log(`${devices.length} device${devices.length > 1 ? 's' : ''} found:`, devices);
+    console.log(`test.js: ${devices.length} device${devices.length > 1 ? 's' : ''} found:`, devices);
   } catch(e) {
-    console.error(e);
+    console.error(`test.js: ${e}`);
   }
-  console.log("Write « FTDI » to access to the driver functions.");
+  console.log("test.js: Write « FTDI » to access to the driver functions.");
 
   // Start REPL
   const cli = repl.start();
   cli.context.FTDI = FTDI;
-})();
+});